feat(canvas): 색상 선택 시 펜 모드 변경 구현 - set 메서드 조건부 분기 수정 - 일단은 toolMode 변경과 함께 setPen 메서드 동시 수행 필수 - (toolMode 만 변경 시 notify 수행 X) - 이제 pen / higlighter 별도 색상 수정 가능
